extract `HTTPClientSpy` to a share scope
